JS Components: Convert AutomatticBylineLogo component to TypeScript (#24944)
* Convert AutomatticBylineLogo logo to TypeScript

* Convert AutomatticBylineLogo tests to TypeScript

* Add more test cases

* Add changelog

* Convert stories to TypeScript

* Fix version

* Fix doc block
