+++ projects/js-packages/components/changelog/update-rna-automattic-byline-logo-to-ts
@@ -1,0 +1,4 @@
+Significance: patch
+Type: changed
+
+JS Components: Converted AutomatticBylineLogo component to TypeScript
